fix bug in mac build process fix bug in resource path management when running a prebuilt Designer app

+++ build.py
@@ -11,7 +11,8 @@ def pyinstall(cmd):
 
 if wx.Platform == "__WXMAC__":
     pyinstall("--onedir --clean -y --windowed -n standalone standalone.py")
-    shutil.rmtree("dist/containerDir")
+    if os.path.exists("dist/containerDir"):
+        shutil.rmtree("dist/containerDir")
     # using --add-data to add a directory includes its contents, not the directory itself, and since apps are
     # directories, then in order to add the app, we need to --add-data a container directory with the app inside of it.
     os.mkdir("dist/containerDir")
